SimSwap/requirements.txt
torch==1.8.0
torchvision==0.9.0
torchaudio==0.8.0
insightface==0.2.1
onnxruntime
moviepy
opencv-python
pillow
numpy
imageio
timm==0.5.4
